Clean up polygon implementations
